refactor(backend/api/db): refactor database operations
the old db_operations.py were modularized
